idsideAI_FIXED/requirements.txt
fastapi>=0.114,<0.116
uvicorn[standard]>=0.30,<0.31
watchfiles>=0.22,<0.25
pydantic==2.7.4
sqlalchemy==2.0.31
sqlmodel==0.0.19
alembic==1.13.2
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
neo4j==5.20.0
python-dotenv==1.0.1
aiosqlite==0.20.0
httpx==0.27.0
stripe==9.12.0
python-docx==1.1.2
prometheus-client==0.20.0
slowapi==0.1.9
jinja2==3.1.4
aiofiles==23.2.1
pytest==8.2.2
pytest-asyncio==0.23.7
requests==2.31.0
reportlab==4.2.2
starlette==0.37.2
pyyaml==6.0.1
pytest
